improve dir structure and fix imports
